Update: Seperate Control Panel as component
