+++ src/features/graphView/GraphView.js
@@ -1,40 +1,39 @@
 import * as d3 from 'd3';
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useEffect, useRef } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { Navigate } from 'react-router-dom';
-import Toggle from 'react-toggle';
 import "react-toggle/style.css";
 import './GraphView.css';
-import { setHostGraphData, setMode, setPortGraphData, setNicknameMapping, resetNicknameMapping } from './graphViewSlice';
-import { Button, Form} from 'react-bootstrap';
+import { setHostGraphData, setPortGraphData } from './graphViewSlice';
+import ControlPanel from './components/ControlPanel';
+import { setIsSimulationStable } from './components/controlPanelSlice';
+
 
 export default function GraphView() {
     const packets = useSelector((state) => state.data.packets);
     const graphRef = useRef(null);
     const hostData = useSelector((state) => state.graphView.hostGraphData);
     const portData = useSelector((state) => state.graphView.portGraphData);
-    const nicknameMapping = useSelector((state) => state.graphView.nicknameMapping);
+    const nicknameMapping = useSelector((state) => state.controlPanel.nicknameMapping);
     const mode = useSelector((state) => state.graphView.mode);
+    const isSimulationStable = useSelector((state) => state.controlPanel.isSimulationStable);
 
     const simulationRef = useRef(null);
     const nodesRef = useRef([]);
     const svgRef = useRef(null); // SVG 요소에 대한 참조를 저장할 ref
     const zoomRef = useRef(null); // zoom behavior에 대한 참조를 저장할 ref
-    const [isSimulationStable, setIsSimulationStable] = useState(false);
+    // const [isSimulationStable, setIsSimulationStable] = useState(false);
     const isSimulationStableRef = useRef(isSimulationStable); 
-    const [isNicknameChangeOpen, setIsNicknameChangeOpen] = useState(false);
-    const [isShowProtocolsOpen, setIsShowProtocolsOpen] = useState(false);
-    const [showL4Protocol, setShowL4Protocol] = useState(false);
-    const [showL7Protocol, setShowL7Protocol] = useState(false);
 
-    const [selectedIP, setSelectedIP] = useState('');
-    const [selectedPort, setSelectedPort] = useState('');
-    const [availablePorts, setAvailablePorts] = useState([]);
-    const [nickname, setNickname] = useState('');
+
+    // const [selectedIP, setSelectedIP] = useState('');
+    // const [selectedPort, setSelectedPort] = useState('');
+    // const [availablePorts, setAvailablePorts] = useState([]);
+    // const [nickname, setNickname] = useState('');
 
 
     function updateIsSimulationStable(value) {
-        setIsSimulationStable(value);
+        dispatch(setIsSimulationStable(value));
         isSimulationStableRef.current = value;
     }
 
@@ -296,7 +295,7 @@ export default function GraphView() {
                 .join("circle")
                 .attr("r", d => sizeScale(d.traffic_volume))
                 .attr("fill", d => colorScale(d.ip_addr))
-                .on("dblclick", resetNodePosition)
+                // .on("dblclick", resetNodePosition)
                 .on("dblclick.zoom", null) // Prevent zoom on double-click on nodes
                 .on("click", (event) => {
                     event.stopPropagation(); // Prevent click from propagating to zoom
@@ -401,19 +400,19 @@ export default function GraphView() {
                     }
                 } else {
                     if (isSimulationStableRef.current) {
-                        updateIsSimulationStable(false);
+                        updateIsSimulationStable(true);
                     }
                 }
             });
 
             simulation.alpha(1).restart();
 
-            function resetNodePosition(event, d) {
-                event.stopPropagation(); // Prevent zoom on double-click
-                d.fx = d.originalX;
-                d.fy = d.originalY;
-                simulation.alpha(1).restart();
-            }
+            // function resetNodePosition(event, d) {
+            //     event.stopPropagation(); // Prevent zoom on double-click
+            //     d.fx = d.originalX;
+            //     d.fy = d.originalY;
+            //     simulation.alpha(1).restart();
+            // }
 
             function zoomed(event) {
                 container.attr("transform", event.transform);
@@ -460,26 +459,6 @@ export default function GraphView() {
             });
     }, [nicknameMapping, mode]);
     
-
-    useEffect(() => {
-        if (mode === 'port' && selectedIP) {
-            const ports = portData.nodes
-                .filter(node => node.ip_addr === selectedIP)
-                .map(node => node.port);
-            const uniquePorts = Array.from(new Set(ports));
-            setAvailablePorts(uniquePorts);
-        } else {
-            setAvailablePorts([]);
-        }
-        setSelectedPort('');
-    }, [selectedIP, mode]);
-    
-    useEffect(() => {
-        setSelectedIP('');
-        setSelectedPort('');
-        setAvailablePorts([]);
-    }, [mode]);
-    
     const getNicknameLabel = (node, mode) => {
         if (mode === 'host') {
             return nicknameMapping[node.ip_addr] || node.ip_addr;
@@ -509,6 +488,7 @@ export default function GraphView() {
                 d.fy = null;
             });
             simulationRef.current.alpha(1).restart();
+            updateIsSimulationStable(false);
         }
 
         if (svgRef.current && zoomRef.current) {
@@ -519,218 +499,10 @@ export default function GraphView() {
         }
     }
 
-    const handleNicknameChange = () => {
-        if (!selectedIP) {
-            alert("Please select an IP address.");
-            return;
-        }
-    
-        if (mode === 'host') {
-            dispatch(setNicknameMapping({ [selectedIP]: nickname }));
-        } else if (mode === 'port') {
-            if (!selectedPort) {
-                alert("Please select a port.");
-                return;
-            }
-            const key = `${selectedIP}:${selectedPort}`;
-            dispatch(setNicknameMapping({ [key]: nickname }));
-        }
-    
-        setNickname('');
-    };
-
-    const handleResetNickname = () => {
-        if (!selectedIP) {
-            alert("Please select an IP address.");
-            return;
-        }
-    
-        if (mode === 'host') {
-            dispatch(resetNicknameMapping(selectedIP));
-        } else if (mode === 'port') {
-            if (!selectedPort) {
-                alert("Please select a port.");
-                return;
-            }
-            const key = `${selectedIP}:${selectedPort}`;
-            dispatch(resetNicknameMapping(key));
-        }
-    };
-
-    const handleResetAllNicknames = () => {
-        dispatch(resetNicknameMapping());
-    };    
-    
-    const handleShowProtocolMenuBar = () => {
-        if (isShowProtocolsOpen) {
-            setShowL4Protocol(false);
-            setShowL7Protocol(false);
-        }
-        setIsShowProtocolsOpen(!isShowProtocolsOpen);
-    };
-
     return (
         packets ? (
             <div style={{ position: "relative", width: "100vw", height: "100vh" }}>
-                {/* Control Panel */}
-                <div style={{ position: 'absolute', top: 70, right: 40, width: "300px", padding: "20px", borderRadius: "15px", border: "1px solid #ccc", backgroundColor: "#fff", boxShadow: "0px 4px 8px rgba(0, 0, 0, 0.1)" }}>
-                    
-                    {/* Port 분리 토글 버튼 */}
-                    <div style={{ cursor: "pointer", display: "flex", justifyContent: "space-between", alignItems: "center", marginBottom: "10px" }}>                        
-                        <Toggle
-                            id='split-toggle'
-                            defaultChecked={mode === 'port'}
-                            onChange={(e) => dispatch(setMode(e.target.checked ? 'port' : 'host'))}
-                        />
-                        <label htmlFor='split-toggle' style={{ marginLeft: '10px',  lineHeight: '0'}}>Split Hosts by Ports</label>
-                    </div>
-
-                    {/* 모든 위치 초기화 버튼 */}
-                    <div style={{ marginBottom: "20px", width: "100%" }}>
-                        <Button onClick={resetAllNodes} style={{ width: "100%" }} disabled={!isSimulationStable}>Reset All Positions</Button>
-                    </div>
-
-                    {/* 구분선 */}
-                    <hr style={{ margin: "20px 0" }} />
-
-                    {/* Nickname Change 메뉴 */}
-                    <div>
-                        <div 
-                            style={{ cursor: "pointer", display: "flex", justifyContent: "space-between", alignItems: "center", marginBottom: "10px" }}
-                            onClick={() => setIsNicknameChangeOpen(!isNicknameChangeOpen)}
-                        >
-                            <strong>Nickname Change</strong>
-                            <span>{isNicknameChangeOpen ? "▼" : "◀"}</span> {/* "◀ ▲ ▶ ▼" */}
-                        </div>
-
-                        {/* Nickname Change 메뉴 아이템 (계층 구조로 보이도록 하되 너비는 줄어들지 않음) */}
-                        {isNicknameChangeOpen && (
-                            <div style={{ marginBottom: "20px", width: "100%" }}>
-                                {/* IP Address 선택 */}
-                                <div style={{ marginBottom: "20px", width: "100%" }}>
-                                    <Form.Label><strong>IP Address:</strong></Form.Label>
-                                    <Form.Select
-                                        className="ip-selector"
-                                        style={{ width: '100%' }}
-                                        onChange={(e) => { setSelectedIP(e.target.value); setSelectedPort(""); }}
-                                        value={selectedIP}
-                                    >
-                                        <option></option>
-                                        {hostData?.nodes.map((opt, index) => (
-                                            <option key={index} value={opt.ip_addr}>{opt.ip_addr}</option>
-                                        ))}
-                                    </Form.Select>
-                                </div>
-
-                                {/* Port Selection (Only visible in 'port' mode) */}
-                                {mode === 'port' && (
-                                    <div style={{ marginBottom: "20px", width: "100%" }}>
-                                        <Form.Label><strong>Port:</strong></Form.Label>
-                                        <Form.Select
-                                            style={{ width: '100%' }}
-                                            onChange={(e) => setSelectedPort(e.target.value)}
-                                            value={selectedPort}
-                                        >
-                                            <option></option>
-                                            {availablePorts.map((port, index) => (
-                                                <option key={index} value={port}>{port}</option>
-                                            ))}
-                                        </Form.Select>
-                                    </div>
-                                )}
-
-                                {/* Nickname Input */}
-                                <div style={{ marginBottom: "20px", width: "100%" }}>
-                                    <Form.Label><strong>Nickname:</strong></Form.Label>
-                                    <Form.Control
-                                        type="text"
-                                        placeholder="Enter nickname"
-                                        value={nickname}
-                                        onChange={(e) => setNickname(e.target.value)}
-                                        style={{ width: '100%' }}
-                                    />
-                                </div>
-
-                                {/* Change 및 Reset 버튼을 한 줄에 배치 */}
-                                <div style={{ marginBottom: "20px", width: "100%", display: "flex", justifyContent: "space-between" }}>
-                                    <Button onClick={handleNicknameChange} style={{ width: "48%" }} disabled={!isSimulationStable}>Change</Button>
-                                    <Button onClick={handleResetNickname} style={{ width: "48%" }} disabled={!isSimulationStable}>Reset</Button>
-                                </div>
-
-                                {/* Reset All Nicknames 버튼 */}
-                                <div style={{ marginBottom: "20px", width: "100%" }}>
-                                    <Button onClick={handleResetAllNicknames} style={{ width: "100%" }} disabled={!isSimulationStable}>Reset All Nicknames</Button>
-                                </div>
-                            </div>
-                        )}
-                    </div>
-
-                    {/* 구분선 */}
-                    <hr style={{ margin: "20px 0" }} />
-
-                    {/* Show Protocols 메뉴바 */}
-                    { mode === 'port' && (
-                    <div>
-                        <div 
-                            style={{ cursor: "pointer", display: "flex", justifyContent: "space-between", alignItems: "center", marginBottom: "10px" }}
-                            onClick={() => handleShowProtocolMenuBar()}
-                        >
-                            <strong>Show Protocols</strong>
-                            <span>{isShowProtocolsOpen ? "▼" : "◀"}</span>
-                        </div>
-
-                        {/* Show Protocols 메뉴 아이템 */}
-                        {isShowProtocolsOpen && (
-                            <div>
-                                <div 
-                                    style={{ cursor: "pointer", display: "flex", justifyContent: "space-between", alignItems: "center", marginBottom: "10px" }}
-                                    onClick={() => setShowL4Protocol(!showL4Protocol)}
-                                >
-                                    <strong>L4 Protocols</strong>
-                                    <span>{showL4Protocol ? "▼" : "◀"}</span>
-                                </div>
-                                {/* L4 Protocol 메뉴 */}
-                                {showL4Protocol && (
-                                    <div style={{ cursor: "pointer", display: "flex", justifyContent: "space-between", alignItems: "center", marginBottom: "10px", borderBottom: "1px dashed #ccc", paddingBottom: "10px" }}>
-                                    {/* <div style={{ marginBottom: "20px", width: "100%" }}> */}
-                                        <Toggle
-                                            id='l4-protocol-toggle'
-                                            checked={showL4Protocol}
-                                            onChange={(e) => setShowL4Protocol(e.target.checked)}
-                                        />
-                                        <label htmlFor='l4-protocol-toggle' style={{ marginLeft: '10px', lineHeight: '0' }}>Show L4 Protocol</label>
-                                    </div>
-                                )}
-                                <div 
-                                    style={{ cursor: "pointer", display: "flex", justifyContent: "space-between", alignItems: "center", marginBottom: "10px" }}
-                                    onClick={() => setShowL7Protocol(!showL7Protocol)}
-                                >
-                                    <strong>L7 Protocols</strong>
-                                    <span>{showL7Protocol ? "▼" : "◀"}</span>
-                                </div>
-                                {/* L7 Protocol 메뉴 */}
-                                {showL7Protocol && (
-                                    <div style={{ cursor: "pointer", display: "flex", justifyContent: "space-between", alignItems: "center", marginBottom: "10px",paddingBottom: "10px" }}>
-                                        <Toggle
-                                            id='l7-protocol-toggle'
-                                            checked={showL7Protocol}
-                                            onChange={(e) => setShowL7Protocol(e.target.checked)}
-                                        />
-                                        <label htmlFor='l7-protocol-toggle' style={{ marginLeft: '10px', lineHeight: '0' }}>Show L7 Protocol</label>
-                                    </div>
-                                )}
-                            </div>
-                        )}
-                    </div>)
-                    
-                    }
-
-                    {/* 구분선 */}
-                    {mode === 'port' && (
-                        <hr style={{ margin: "20px 0" }} />
-                    )}
-
-                </div>
+                <ControlPanel resetAllNodes={resetAllNodes}/>
 
                 {/* Graph Area */}
                 <svg ref={graphRef} style={{ width: '100%', height: '100%' }} />
